requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
azure-cosmos==4.5.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
et-xmlfile==1.1.0
Flask==2.2.2
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.2
openai==0.26.5
openpyxl==3.1.0
pandas==1.5.3
pandas-stubs==1.5.3.230203
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
toml==0.10.2
tqdm==4.64.1
types-pytz==2022.7.1.0
urllib3==1.26.14
Werkzeug==2.2.2
yarl==1.8.2